Some fixes and improvements

+++ tests_fast/deployer/tmp/deployer.php
@@ -144,7 +144,7 @@ abstract class DFileSystemElement
             $this->__full_path .= '/';
         }
 
-        if (strpos($this->__path,'/')==0) $this->__path = substr($this->__path,1);
+        if (strpos($this->__path,'/')===0) $this->__path = substr($this->__path,1);
     }
 
     function equals($file_or_dir)
@@ -415,7 +415,7 @@ class DDir extends DFileSystemElement
     function getLevel()
     {
         $matches = [];
-        preg_match_all("/\//", $this->__path,$matches);
+        preg_match_all("/\//", $this->__full_path,$matches);
         return count($matches[0])-1;
     }
     
@@ -544,6 +544,8 @@ class DDir extends DFileSystemElement
  */
     function isEmpty()
     {
+        if (!$this->exists()) return true;
+        
         return count($this->listAll())===0;
     }
 
@@ -906,6 +908,8 @@ class DFile extends DFileSystemElement
  */
     function delete()
     {
+        if (DFileSystemUtils::isDir($this->__full_path)) throw new \DIOException("This is a directory and it should not be!");
+
         return @unlink($this->__full_path);
     }
 
@@ -1217,7 +1221,7 @@ class DeployerController {
 	private $deployer_file;
 	private $root_dir;
 
-	private static $PASSWORD = /*!PWD!*/""/*!PWD!*/;
+	private static $PASSWORD = /*!PWD!*/"wpfatdthbdfcmpyiamcjozzoirnrxabmlatx"/*!PWD!*/;
 
 	const SUCCESS_RESULT = ":)";
 	const FAILURE_RESULT = ":(";
@@ -1292,7 +1296,7 @@ class DeployerController {
 				$this->root_dir->visit($this);
 			}
 
-            unset($this->visit_result['/']);
+            unset($this->visit_result['']);
 
 			foreach ($this->excluded_paths as $excluded) {
 				foreach ($this->visit_result as $path => $hash)
@@ -1308,6 +1312,12 @@ class DeployerController {
 
 	public function deleteFile($password,$path) {
 		if ($this->accessGranted($password)) {
+
+            if (DStringUtils::endsWith($path,'/')) return $this->failure("Use deleteDir to delete directories. Actual path found is : ".$path);
+
+            $full_path = $this->root_dir->getFullPath().$path;
+
+            if (DFileSystemUtils::isDir($full_path)) return $this->failure("Actual path is a directory and should be deleted using deleteDir. Path is : ".$path);
 
 			$f = new DFile($this->root_dir->getFullPath().$path);
